+++ src/DeviceLayout.jl
@@ -18,6 +18,8 @@ import Unitful: Length, LengthUnits, DimensionlessQuantity, NoUnits, DimensionEr
 import Unitful: ustrip, unit, inch
 Unitful.@derived_dimension InverseLength inv(Unitful.𝐋)
 
+import SpatialIndexing
+
 function render! end
 export render!
 
@@ -194,6 +196,7 @@ coordinatetype(::AbstractArray{S}) where {T, S <: AbstractGeometry{T}} = T
 coordinatetype(iterable) = promote_type(coordinatetype.(iterable)...)
 coordinatetype(::Point{T}) where {T} = T
 coordinatetype(::Type{Point{T}}) where {T} = T
+coordinatetype(::Pair{<:AbstractGeometry{T}}) where {T} = T
 
 # Entity interface
 include("entities.jl")
@@ -203,8 +206,10 @@ export GeometryEntity,
     center,
     centered,
     coordinatetype,
+    findbox,
     footprint,
     halo,
+    mbr_spatial_index,
     offset,
     to_polygons,
     lowerleft,
@@ -379,8 +384,10 @@ import .Polygons:
     clip,
     cliptree,
     difference2d,
+    difference2d_layerwise,
     gridpoints_in_polygon,
     intersect2d,
+    intersect2d_layerwise,
     offset,
     perimeter,
     points,
@@ -389,7 +396,9 @@ import .Polygons:
     sweep_poly,
     unfold,
     union2d,
-    xor2d
+    union2d_layerwise,
+    xor2d,
+    xor2d_layerwise
 export Polygons,
     Polygon,
     Ellipse,
@@ -405,8 +414,10 @@ export Polygons,
     clip,
     cliptree,
     difference2d,
+    difference2d_layerwise,
     gridpoints_in_polygon,
     intersect2d,
+    intersect2d_layerwise,
     offset,
     perimeter,
     points,
@@ -415,7 +426,9 @@ export Polygons,
     sweep_poly,
     unfold,
     union2d,
-    xor2d
+    union2d_layerwise,
+    xor2d,
+    xor2d_layerwise
 
 include("align.jl")
 using .Align